[Refactor ⚙️] 폴더 분리
